refactor: 重构 command 模块并修复所有 clippy 警告
- 将 lib.rs 中的 command 拆分到 src/command/ 目录下
- 修复 14 个 clippy 警告：
  - unused_imports: 删除未使用的导入
  - manual_is_multiple_of: 使用 is_multiple_of()
  - empty_line_after_doc_comments: 删除文档注释后的空行
  - redundant_pattern_matching: 使用 is_some() 替代 if let Some(_)
  - collapsible_if: 合并嵌套的 if 语句
  - needless_return: 移除不必要的 return
  - double_ended_iterator_last: 使用 next_back() 替代 last()
  - useless_format: 使用 to_string() 替代 format!("...")
  - unnecessary_cast: 移除多余的类型转换
  - too_many_arguments: 添加 #[allow] 注解
  - iter_cloned_collect: 使用 to_vec() 替代 iter().copied().collect()
  - single_match: 使用 if let 替代 match

+++ src-tauri/src/services/sync_manager.rs
@@ -273,7 +273,7 @@ impl SyncManager {
             let _ = crate::services::folder_service::update_stats(
                 &self.db,
                 folder_id,
-                email_count as i32,
+                email_count,
                 unread_count,
             )
             .await;
